refactor: remove old ways of creating non-EVM accounts
