+++ app/_locales/en/messages.json
@@ -1880,15 +1880,6 @@
   "bridgeConfirmTwoTransactions": {
     "message": "You'll need to confirm 2 transactions on your hardware wallet:"
   },
-  "bridgeCreateSolanaAccount": {
-    "message": "Create Solana account"
-  },
-  "bridgeCreateSolanaAccountDescription": {
-    "message": "To swap to the Solana network, you need an account and receiving address."
-  },
-  "bridgeCreateSolanaAccountTitle": {
-    "message": "You'll need a Solana account first."
-  },
   "bridgeDetailsTitle": {
     "message": "Bridge details",
     "description": "Title for the modal showing details about a bridge transaction."
@@ -2163,9 +2154,6 @@
   },
   "bridgeSendLoading": {
     "message": "Sending from source chain"
-  },
-  "bridgeSolanaAccountCreated": {
-    "message": "Solana account created"
   },
   "bridgeStatusComplete": {
     "message": "Complete",
