+++ app/_locales/es_419/messages.json
@@ -1178,15 +1178,6 @@
   "bridgeConfirmTwoTransactions": {
     "message": "Deberá confirmar 2 transacciones en su monedero físico:"
   },
-  "bridgeCreateSolanaAccount": {
-    "message": "Crear cuenta de Solana"
-  },
-  "bridgeCreateSolanaAccountDescription": {
-    "message": "Para cambiar a la red Solana, necesita una cuenta y una dirección receptora."
-  },
-  "bridgeCreateSolanaAccountTitle": {
-    "message": "Primero necesitará una cuenta de Solana."
-  },
   "bridgeDetailsTitle": {
     "message": "Detalles del puenteo",
     "description": "Title for the modal showing details about a bridge transaction."
@@ -1461,9 +1452,6 @@
   },
   "bridgeSendLoading": {
     "message": "Enviando desde la cadena de origen"
-  },
-  "bridgeSolanaAccountCreated": {
-    "message": "Se creó la cuenta de Solana"
   },
   "bridgeStatusComplete": {
     "message": "Completada",
