+++ app/_locales/fr/messages.json
@@ -1172,15 +1172,6 @@
   "bridgeConfirmTwoTransactions": {
     "message": "Vous devrez confirmer 2 transactions sur votre portefeuille matériel :"
   },
-  "bridgeCreateSolanaAccount": {
-    "message": "Créer un compte Solana"
-  },
-  "bridgeCreateSolanaAccountDescription": {
-    "message": "Pour échanger avec le réseau Solana, vous avez besoin d’un compte et d’une adresse de réception."
-  },
-  "bridgeCreateSolanaAccountTitle": {
-    "message": "Vous devez d’abord avoir un compte Solana."
-  },
   "bridgeDetailsTitle": {
     "message": "Détails concernant la passerelle",
     "description": "Title for the modal showing details about a bridge transaction."
@@ -1455,9 +1446,6 @@
   },
   "bridgeSendLoading": {
     "message": "Envoi depuis la chaîne source"
-  },
-  "bridgeSolanaAccountCreated": {
-    "message": "Compte Solana créé"
   },
   "bridgeStatusComplete": {
     "message": "Effectuée",
