+++ app/_locales/tl/messages.json
@@ -1178,15 +1178,6 @@
   "bridgeConfirmTwoTransactions": {
     "message": "Kakailanganin mong kumpirmahin ang 2 transaksyon sa iyong hardware wallet:"
   },
-  "bridgeCreateSolanaAccount": {
-    "message": "Gumawa ng Solana account"
-  },
-  "bridgeCreateSolanaAccountDescription": {
-    "message": "Para mag-swap papunta sa Solana network, kailangan mo ng account at tatanggap na address."
-  },
-  "bridgeCreateSolanaAccountTitle": {
-    "message": "Kailangan mo muna ng Solana account."
-  },
   "bridgeDetailsTitle": {
     "message": "Mga detalye ng bridge",
     "description": "Title for the modal showing details about a bridge transaction."
@@ -1461,9 +1452,6 @@
   },
   "bridgeSendLoading": {
     "message": "Ipinapadala mula sa chain ng pinagmulan"
-  },
-  "bridgeSolanaAccountCreated": {
-    "message": "Nakagawa ng Solana account"
   },
   "bridgeStatusComplete": {
     "message": "Kumpleto na",
