+++ shared/lib/accounts/accounts.ts
@@ -1,81 +1,5 @@
-import { InternalAccount } from '@metamask/keyring-internal-api';
-import { CaipChainId } from '@metamask/utils';
-import { DiscoveredAccount, KeyringAccount } from '@metamask/keyring-api';
-import {
-  KeyringInternalSnapClient,
-  KeyringInternalSnapClientMessenger,
-} from '@metamask/keyring-internal-snap-client';
-import {
-  SnapKeyring,
-  SnapKeyringInternalOptions,
-} from '@metamask/eth-snap-keyring';
-import { Messenger } from '@metamask/messenger';
-import { SnapId } from '@metamask/snaps-sdk';
-import { SnapControllerHandleRequestAction } from '@metamask/snaps-controllers';
-import { MultichainNetworks } from '../../constants/multichain/networks';
-import { createSentryError } from '../error';
-import { captureException } from '../sentry';
+import type { InternalAccount } from '@metamask/keyring-internal-api';
 import { HardwareDeviceNames } from '../../constants/hardware-wallets';
-import { BITCOIN_WALLET_SNAP_ID } from './bitcoin-wallet-snap';
-import { SOLANA_WALLET_SNAP_ID } from './solana-wallet-snap';
-import { TRON_WALLET_SNAP_ID } from './tron-wallet-snap';
-
-/**
- * Supported non-EVM Snaps.
- */
-// TODO: Fix in https://github.com/MetaMask/metamask-extension/issues/31860
-// eslint-disable-next-line @typescript-eslint/naming-convention
-type SUPPORTED_WALLET_SNAP_ID =
-  | typeof SOLANA_WALLET_SNAP_ID
-  | typeof BITCOIN_WALLET_SNAP_ID
-  | typeof TRON_WALLET_SNAP_ID;
-
-export type SnapAccountNameOptions = {
-  chainId?: CaipChainId;
-};
-
-/**
- * Get the next available Snap account name for a supported non-EVM Snap.
- *
- * @param getNextAvailableAccountName - Callback to get the next available account name.
- * @param snapId - Snap ID.
- * @param options - Options for this account name.
- * @param options.chainId - Chain ID for this account if available.
- * @returns
- */
-export async function getNextAvailableSnapAccountName(
-  getNextAvailableAccountName: () => Promise<string>,
-  snapId: SUPPORTED_WALLET_SNAP_ID,
-  { chainId }: SnapAccountNameOptions = {},
-) {
-  const defaultSnapAccountName = await getNextAvailableAccountName();
-
-  // FIXME: This is a temporary workaround to suggest a different account name for a first party snap.
-  const accountNumber = defaultSnapAccountName.trim().split(' ').pop();
-
-  switch (snapId) {
-    case BITCOIN_WALLET_SNAP_ID: {
-      if (chainId === MultichainNetworks.BITCOIN_TESTNET) {
-        return `Bitcoin Testnet Account ${accountNumber}`;
-      }
-      if (chainId === MultichainNetworks.BITCOIN_SIGNET) {
-        return `Bitcoin Signet Account ${accountNumber}`;
-      }
-      return `Bitcoin Account ${accountNumber}`;
-    }
-    case SOLANA_WALLET_SNAP_ID: {
-      // Solana accounts should have in their scope the 3 networks
-      // mainnet, testnet, and devnet. Therefore, we can use this name
-      // for all 3 networks.
-      return `Solana Account ${accountNumber}`;
-    }
-    case TRON_WALLET_SNAP_ID: {
-      return `Tron Account ${accountNumber}`;
-    }
-    default:
-      return defaultSnapAccountName;
-  }
-}
 
 export function isHardwareAccount(account: InternalAccount): boolean {
   try {
@@ -85,149 +9,5 @@ export function isHardwareAccount(account: InternalAccount): boolean {
     );
   } catch {
     return false;
-  }
-}
-
-export type CreateAccountSnapOptions = {
-  scope?: CaipChainId;
-  derivationPath?: DiscoveredAccount['derivationPath'];
-  entropySource?: string;
-  accountNameSuggestion?: string;
-  synchronize?: boolean;
-};
-
-export type WalletSnapClient = {
-  getSnapId(): SnapId;
-
-  createAccount(
-    options: CreateAccountSnapOptions,
-    internalOptions?: SnapKeyringInternalOptions,
-  ): Promise<KeyringAccount>;
-
-  getNextAvailableAccountName(
-    options?: SnapAccountNameOptions,
-  ): Promise<string>;
-};
-
-export type MultichainWalletSnapClientMessenger = Messenger<
-  'MultichainWalletSnapClient',
-  SnapControllerHandleRequestAction,
-  never
->;
-
-export class MultichainWalletSnapClient implements WalletSnapClient {
-  readonly #snapId: SUPPORTED_WALLET_SNAP_ID;
-
-  readonly #snapKeyring: SnapKeyring;
-
-  readonly #client: KeyringInternalSnapClient;
-
-  constructor(
-    snapId: SUPPORTED_WALLET_SNAP_ID,
-    snapKeyring: SnapKeyring,
-    messenger: MultichainWalletSnapClientMessenger,
-  ) {
-    this.#snapId = snapId;
-    this.#snapKeyring = snapKeyring;
-
-    const clientMessenger: KeyringInternalSnapClientMessenger = new Messenger({
-      namespace: 'KeyringInternalSnapClient',
-      parent: messenger,
-    });
-
-    messenger.delegate({
-      messenger: clientMessenger,
-      actions: ['SnapController:handleRequest'],
-    });
-
-    this.#client = new KeyringInternalSnapClient({
-      snapId,
-      messenger: clientMessenger,
-    });
-  }
-
-  getSnapId(): SnapId {
-    return this.#snapId;
-  }
-
-  async createAccount(
-    { accountNameSuggestion, ...options }: CreateAccountSnapOptions,
-    internalOptions?: SnapKeyringInternalOptions,
-  ): Promise<KeyringAccount> {
-    // Automatically name the account if not provided.
-    const autoAccountNameSuggestion = accountNameSuggestion
-      ? { accountNameSuggestion }
-      : { accountNameSuggestion: await this.getNextAvailableAccountName() };
-
-    // TODO: Use `withKeyring` instead of using the keyring directly.
-    return await this.#snapKeyring.createAccount(
-      this.#snapId,
-      {
-        ...options,
-        // TODO: Stop forwarding the account name to the Snap, we should make this
-        // an internal option at some point.
-        ...autoAccountNameSuggestion,
-      },
-      internalOptions,
-    );
-  }
-
-  async getNextAvailableAccountName(
-    options: SnapAccountNameOptions = {},
-  ): Promise<string> {
-    return getNextAvailableSnapAccountName(
-      async () => {
-        return '';
-      },
-      this.#snapId,
-      options,
-    );
-  }
-
-  async discoverAccounts(
-    entropySource: string,
-    scope: CaipChainId,
-  ): Promise<KeyringAccount[]> {
-    const accounts: KeyringAccount[] = [];
-
-    for (let index = 0; ; index++) {
-      const discovered = await this.#client.discoverAccounts(
-        [scope],
-        entropySource,
-        index,
-      );
-
-      // We stop discovering accounts if none got discovered for that index.
-      if (discovered.length === 0) {
-        break;
-      }
-
-      // NOTE: We are doing this sequentially mainly to avoid race-conditions with the
-      // account naming logic.
-      for (const { derivationPath } of discovered) {
-        try {
-          const options: CreateAccountSnapOptions = {
-            scope,
-            derivationPath,
-            entropySource,
-            synchronize: true,
-          };
-
-          const account = await this.createAccount(options, {
-            displayConfirmation: false,
-            displayAccountNameSuggestion: false,
-            setSelectedAccount: false,
-          });
-          accounts.push(account);
-        } catch (error) {
-          // Still logging this one to sentry as this is a fairly new process for account discovery.
-          captureException(
-            createSentryError('Unable to create discovered account', error),
-          );
-        }
-      }
-    }
-
-    return accounts;
   }
 }
